wip namespace events in notifications
